inspection error fixes

+++ src/main/kotlin/environment/errorsCatchable/nonBlockError.kt
@@ -5,5 +5,5 @@ import data.registers.intelNames
 import registers
 
 fun nonBlockError(errorType: ErrorType) {
-	registers.write(intelNames[IntelRegisters.ESF], errorType.code.toLong())
+	registers.write(intelNames[IntelRegisters.ESF], errorType.code)
 }
